+++ lib/test/clojure_lsp/shared_test.clj
@@ -16,13 +16,13 @@
     (is (= {:a {:b [1 2 4 3] :c 3}} (shared/deep-merge {:a {:b [1 2]}} {:a {:b #{3 4} :c 3}})))))
 
 (deftest external-filename?
-  (is (not (shared/external-filename? "/some/project/src/a.clj" #{"/some/project/src"})))
-  (is (not (shared/external-filename? "/some/project/src/a.clj" #{})))
-  (is (shared/external-filename? "/some/project/src/a.clj" #{"/some/project/src/b.clj"}))
-  (is (shared/external-filename? "/some/place/file.jar:some/path/to/file.clj" #{"/some/project/src/a.clj"}))
-  (is (shared/external-filename? "/some/place/file.jar:some/path/to/file.clj" #{}))
-  (is (shared/external-filename? "/some/place/file.jar:some/path/to/file.clj" #{"/some/place/file.jar:some/path"}))
-  (is (shared/external-filename? "/some/user/.emacs.d/.local/etc/workspace/.cache/something.cljc" #{"/some/place/file.clj"})))
+  (is (not (shared/external-filename? (h/file-path "/some/project/src/a.clj") #{(h/file-path "/some/project/src")})))
+  (is (not (shared/external-filename? (h/file-path "/some/project/src/a.clj") #{})))
+  (is (shared/external-filename? (h/file-path "/some/project/src/a.clj") #{(h/file-path "/some/project/src/b.clj")}))
+  (is (shared/external-filename? (h/file-path "/some/place/file.jar:some/path/to/file.clj") #{(h/file-path "/some/project/src/a.clj")}))
+  (is (shared/external-filename? (h/file-path "/some/place/file.jar:some/path/to/file.clj") #{}))
+  (is (shared/external-filename? (h/file-path "/some/place/file.jar:some/path/to/file.clj") #{(h/file-path "/some/place/file.jar:some/path")}))
+  (is (shared/external-filename? (h/file-path "/some/user/.emacs.d/.local/etc/workspace/.cache/something.cljc") #{(h/file-path "/some/place/file.clj")})))
 
 (deftest uri->filename
   (testing "should decode special characters in file URI"
@@ -187,6 +187,15 @@
          (shared/namespace+source-path->filename "some.cool-ns" (h/file-path "/project/test") :clj)))
   (is (= (h/file-path "/project/test/some/cool_ns.clj")
          (shared/namespace+source-path->filename "some.cool-ns" (h/file-path "/project/test/") :clj))))
+
+(deftest uri->source-paths
+  (is (= [(h/file-path "/dir/project/src")]
+         (shared/uri->source-paths (h/file-uri "file:///dir/project/src/clj/a/b.clj")
+                                   [(h/file-path "/dir/project/test") (h/file-path "/dir/project/src")])))
+  (testing "one source-path is a prefix of another"
+    (is (= [(h/file-path "/dir/project/src/cljs")]
+           (shared/uri->source-paths (h/file-uri "file:///dir/project/src/cljs/a/b.clj")
+                                     [(h/file-path "/dir/project/src/clj") (h/file-path "/dir/project/src/cljs")])))))
 
 (deftest jar-file?-test
   (is (= false (shared/jar-file? "")))