+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - General
   - Bump clj-depend to `0.10.0`.
+  - Fix renaming ns when one source path is a prefix of another (e.g. `src/clj` and `src/cljs`) #1805
 
 - Editor
   - Use scope location instead of name location for folding feature.